align_depth_to_infra2 enabled, pointcloud and align_depth filters to own files
